Change type of pricelist, brandId and priority to Integer.

+++ pricing-ms/src/main/java/com/ecommerce/pricing/application/port/in/GetProductPriceUseCase.java
@@ -6,5 +6,5 @@ import java.util.Optional;
 import com.ecommerce.pricing.application.dto.PricesResponseDTO;
 
 public interface GetProductPriceUseCase {
-    Optional<PricesResponseDTO> getProductPrice(Long brandId, Long productId, LocalDateTime applicationDate);
+    Optional<PricesResponseDTO> getProductPrice(Integer brandId, Long productId, LocalDateTime applicationDate);
 }
